+++ file2
@@ -1,0 +1,1 @@
+World
