+++ file3
@@ -1,0 +1,1 @@
+lol
